Simplificando el codigo
